feat: Refactor DesktopNav and MobileNav components; extract styled components to style.ts for better organization
